Fixed ImGUI being a float intead of an int

+++ Game.h
@@ -104,7 +104,7 @@ private:
 	Microsoft::WRL::ComPtr<ID3D11RenderTargetView> ppRTV; // For rendering
 	Microsoft::WRL::ComPtr<ID3D11ShaderResourceView> ppSRV; // For sampling
 
-	float blurAmount = 0;
+	int blurAmount = 0;
